Move validate_sample_ids to Samples controller and rename it to validate_access_to_sample_ids
