+++ app/assets/src/api/access_control.js
@@ -1,0 +1,6 @@
+import { postWithCSRF } from "./core";
+
+export const validateSampleIds = sampleIds =>
+  postWithCSRF("/samples/validate_access_to_sample_ids", {
+    sampleIds: sampleIds,
+  });
