+++ app/assets/src/api/bulk_downloads.js
@@ -16,8 +16,3 @@ export const getBulkDownload = bulkDownloadId =>
 
 export const getPresignedOutputUrl = bulkDownloadId =>
   get(`/bulk_downloads/${bulkDownloadId}/presigned_output_url`);
-
-export const validateSampleIds = sampleIds =>
-  postWithCSRF("/bulk_downloads/validate_sample_ids", {
-    sampleIds: sampleIds,
-  });
