+++ spec/controllers/bulk_downloads_controller_spec.rb
@@ -12,103 +12,6 @@ RSpec.describe BulkDownloadsController, type: :controller do
       @project = create(:project, users: [@joe], name: "Test Project")
     end
 
-    describe "on POST #validate" do
-      before do
-        AppConfigHelper.set_app_config(AppConfig::MAX_SAMPLES_BULK_DOWNLOAD, 100)
-        @admin_project = create(:project, users: [@admin])
-      end
-
-      let(:good_sample_one) do
-        create(:sample, project: @project, name: "Test Sample One", pipeline_runs_data: [{ finalized: 1, job_status: PipelineRun::STATUS_CHECKED }])
-      end
-
-      let(:good_sample_two) do
-        create(:sample, project: @project, name: "Test Sample Two", pipeline_runs_data: [{ finalized: 1, job_status: PipelineRun::STATUS_CHECKED }])
-      end
-
-      let(:in_progress_sample) do
-        create(:sample, project: @project, name: "In Progress Sample", pipeline_runs_data: [{ finalized: 0, job_status: PipelineRun::STATUS_RUNNING }])
-      end
-
-      let(:failed_sample) do
-        create(:sample, project: @project, name: "Failed Sample", pipeline_runs_data: [{ finalized: 1, job_status: PipelineRun::STATUS_FAILED }])
-      end
-
-      let(:different_owner_sample) do
-        create(:sample, project: @admin_project, name: "Admin Sample", pipeline_runs_data: [{ finalized: 1, job_status: PipelineRun::STATUS_CHECKED }])
-      end
-
-      it "should validate successful samples owned by user" do
-        validate_params = {
-          sampleIds: [good_sample_one, good_sample_two],
-        }
-
-        post :validate_sample_ids, params: validate_params
-
-        expect(response).to have_http_status(200)
-        json_response = JSON.parse(response.body)
-
-        expect(json_response).not_to eq(nil)
-        expect(json_response["validSampleIds"]).to include(good_sample_one.id)
-        expect(json_response["validSampleIds"]).to include(good_sample_two.id)
-        expect(json_response["invalidSampleNames"]).to be_empty
-        expect(json_response["error"]).to be_nil
-      end
-
-      it "should filter out samples in progress" do
-        validate_params = {
-          sampleIds: [good_sample_one, good_sample_two, in_progress_sample],
-        }
-
-        post :validate_sample_ids, params: validate_params
-
-        expect(response).to have_http_status(200)
-        json_response = JSON.parse(response.body)
-
-        expect(json_response).not_to eq(nil)
-        expect(json_response["validSampleIds"]).to include(good_sample_one.id)
-        expect(json_response["validSampleIds"]).to include(good_sample_two.id)
-        expect(json_response["invalidSampleNames"]).to include(in_progress_sample.name)
-        expect(json_response["error"]).to be_nil
-      end
-
-      it "should filter out failed samples" do
-        validate_params = {
-          sampleIds: [good_sample_one, good_sample_two, failed_sample],
-        }
-
-        post :validate_sample_ids, params: validate_params
-
-        expect(response).to have_http_status(200)
-        json_response = JSON.parse(response.body)
-
-        expect(json_response).not_to eq(nil)
-        expect(json_response["validSampleIds"]).to include(good_sample_one.id)
-        expect(json_response["validSampleIds"]).to include(good_sample_two.id)
-        expect(json_response["invalidSampleNames"]).to include(failed_sample.name)
-        expect(json_response["error"]).to be_nil
-      end
-
-      it "should silently filter out samples the user cannot view" do
-        validate_params = {
-          sampleIds: [good_sample_one, good_sample_two, different_owner_sample],
-        }
-
-        expect(Rails.logger).to receive(:warn).with(/SampleAccessValidationImproperAccessEvent: User with id #{@joe.id} made a request for samples they don't have access to: \[#{different_owner_sample.id}\]/)
-
-        post :validate_sample_ids, params: validate_params
-
-        expect(response).to have_http_status(200)
-        json_response = JSON.parse(response.body)
-
-        expect(json_response).not_to eq(nil)
-        expect(json_response["validSampleIds"]).to include(good_sample_one.id)
-        expect(json_response["validSampleIds"]).to include(good_sample_two.id)
-        expect(json_response["invalidSampleNames"]).to be_empty
-        expect(json_response["error"]).to be_nil
-      end
-    end
-
     describe "POST #create" do
       before do
         AppConfigHelper.set_app_config(AppConfig::MAX_SAMPLES_BULK_DOWNLOAD, 100)
@@ -582,41 +485,6 @@ RSpec.describe BulkDownloadsController, type: :controller do
         expect(bulk_downloads[0]["user_name"]).to eq("Joe")
         expect(bulk_downloads[0]["execution_type"]).to eq("ecs")
         expect(bulk_downloads[0]["log_url"]).to match("MOCK_TASK_ARN")
-      end
-    end
-
-    describe "POST #validate" do
-      before do
-        AppConfigHelper.set_app_config(AppConfig::MAX_SAMPLES_BULK_DOWNLOAD, 100)
-        @joe_project = create(:project, users: [@joe])
-      end
-
-      let(:good_sample_one) do
-        create(:sample, project: @project, name: "Test Sample One", pipeline_runs_data: [{ finalized: 1, job_status: PipelineRun::STATUS_CHECKED }])
-      end
-
-      let(:good_sample_two) do
-        create(:sample, project: @project, name: "Test Sample Two", pipeline_runs_data: [{ finalized: 1, job_status: PipelineRun::STATUS_CHECKED }])
-      end
-
-      let(:different_owner_sample) do
-        create(:sample, project: @joe_project, name: "Joe Sample", pipeline_runs_data: [{ finalized: 1, job_status: PipelineRun::STATUS_CHECKED }])
-      end
-
-      it "should validate samples owned by any user" do
-        validate_params = {
-          sampleIds: [different_owner_sample],
-        }
-
-        post :validate_sample_ids, params: validate_params
-
-        expect(response).to have_http_status(200)
-        json_response = JSON.parse(response.body)
-
-        expect(json_response).not_to eq(nil)
-        expect(json_response["validSampleIds"]).to include(different_owner_sample.id)
-        expect(json_response["invalidSampleNames"]).to be_empty
-        expect(json_response["error"]).to be_nil
       end
     end
